% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{ee.gee.onlymean}
\alias{ee.gee.onlymean}
\title{Estimating equation of marginal mean for GEE without missing}
\usage{
ee.gee.onlymean(y,x,r,id,beta,ro,phi,dist,corstr)
}
\arguments{
\item{y}{A vector containing outcomes.}

\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records. The default setup is that all data are observed.}

\item{id}{A vector indicating subject id.}

\item{beta}{A plug-in estimator solved by an external estimation procedure, such as GEE.}

\item{ro}{A correlation coefficients obtained from an external estimation procedure, such as GEE.}

\item{phi}{An over-dispersion parameter obtained from an external estimation procedure, such as GEE.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{corstr}{A condidate correlation structure. It can be "independence","exchangeable", and "ar1".}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
Calculate estimating equation from GEE in ELCIC. This estimating equation is used for marginal mean selection.
}
\note{
corstr should be prespecified.
}
\examples{
## tests
# load data
data(geesimdata)
x<-geesimdata$x
y<-geesimdata$y
id<-geesimdata$id
corstr<-"exchangeable"
dist<-"poisson"
# obtain the estimates
library(geepack)
fit<-geeglm(y~x-1,data=geesimdata,family =dist,id=id,corstr = corstr)
beta<-fit$coefficients
ro<-unlist(summary(fit)$corr[1])
phi<-unlist(summary(fit)$dispersion[1])
r<-rep(1,nrow(x))
ee.matrix<-ee.gee.onlymean(y,x,r,id,beta,ro,phi,dist,corstr)
apply(ee.matrix,1,mean)

}
