% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{ee.wgee.onlymean}
\alias{ee.wgee.onlymean}
\title{Estimating equation for marginal mean under weighted GEE(WGEE) with data missing at random}
\usage{
ee.wgee.onlymean(y,x,r,pi,id,time,beta,ro,phi,dist,corstr)
}
\arguments{
\item{y}{A vector containing outcomes. use NA to indicate missing outcomes.}

\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records.}

\item{pi}{A vector containing observing probabilities across all observations.}

\item{id}{A vector indicating subject id.}

\item{time}{The number of observations for each subject}

\item{beta}{A plug-in estimator solved by an external estimation procedure, such as WGEE.}

\item{ro}{A correlation coefficients obtained from an external estimation procedure, such as WGEE.}

\item{phi}{An over-dispersion parameter obtained from an external estimation procedure, such as GEE.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{corstr}{A condidate correlation structure. It can be "independence","exchangeable", and "ar1".}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
Calculate estimating equation from WGEE in ELCIC. This estimating equation is used for marginal mean selection.
}
\note{
corstr should be prespecified.
}
\examples{
## tests
# load data
data(wgeesimdata)
library(wgeesel)
data_wgee<-data.frame(do.call(cbind,wgeesimdata))
corstr<-"exchangeable"
dist<-"binomial"
id<-data_wgee$id
# obtain the estimates
fit<-wgee(y~x1+x2+x3,data_wgee,id,family=dist,corstr =corstr,
     scale = NULL,mismodel =obs_ind~x_mis1)
beta<-as.vector(summary(fit)$beta)
ro<-summary(fit)$corr
phi<-summary(fit)$phi
#calculate observing probabilies for all observations
gamma<-as.vector(summary(fit$mis_fit)$coefficients[,1])
x_mis<-wgeesimdata$x_mis
pi<-prob.obs(x_mis,gamma,id,time=3)
wgee.matrix<-ee.wgee.onlymean(y=wgeesimdata$y,x=wgeesimdata$x,r=wgeesimdata$obs_ind,
pi=pi,id=wgeesimdata$id,time=3,beta=beta,ro=ro,phi=phi,dist=dist,corstr=corstr)
apply(wgee.matrix,1,mean)
}
