% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.wgee.single}
\alias{ELCIC.wgee.single}
\title{Calculate ELCIC for a given candidate model under the WGEE framework for longitudinal data with drop-out and the assumption of missing at random}
\usage{
ELCIC.wgee.single(x,y,x_mis,r,id,time,index.var=NULL,
      name.var=NULL,dist,corstr,joints=TRUE,lag=1)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept if the intercept is considered in the candidate model. The covariate matrix should be complete. See more in Details section.}

\item{y}{A vector containing outcomes. Use NA to indicate missing outcomes.}

\item{x_mis}{A matrix containing covariates for the missing data model. The first column should be all ones corresponding to the intercept. This covariate matrix should be complete and all observed. See more in Details section.}

\item{r}{A vector indicating the observation of outcomes: 1 for an observed record, and 0 for an unobserved record.}

\item{id}{A vector indicating subject id.}

\item{time}{The number of observations in total for each subject}

\item{index.var}{A vector containing index corresponding to the covariates in the candidate model. See more in Details section.}

\item{name.var}{A vector containing names of covariates in the candidate model. The names should be a subset of the column names of the x matrix. See more in Details section.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson", and "binomial".}

\item{corstr}{A condidate correlation structure. It can be "independence","exchangeable", and "ar1".}

\item{joints}{A logic value for joint selection of marginal mean and "working" correlation structures. The default is TRUE. See more in Details section.}

\item{lag}{A numeric value indicating lag-response involved in the missing data model. It can be one of 0, 1, and 2. The default is 1.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
The function \code{\link{ELCIC.wgee.single}} to calculate ELCIC for a given candidate mean model with a specified "working" correlation structure. It is able to simultaneously evaluate the marginal mean and "working" correlation structures. The longutindal data has dropout missingness with the assumption of missing at random.
}
\details{
The covariate matrix "x" should be complete. If missing data are present in "x", the elements in covariate vector will be replaced by zeros for individuals who have missing covariates.

The argument "x_mis" includes all covariates to fit the missing data model. It does not contains a lag variable based on the outcome y. The argument "lag" in this function will automatically add lag-response variables to indicate the data with missing at random????.

Either of the arguments, "index.var" and "name.var", is used to identify the candidate mean model. If both arguments are provided, only the argument "name.var" will be used.

When the argument "joints" is TRUE, \code{\link{ELCIC.wgee.single}} will calculate ELCIC based on the function \code{\link{lambda.find.wgee}} and \code{\link{ee.wgee}}, which involve estimating equations for both marginal mean and correlation coefficients. When the argument "joints" is FALSE, \code{\link{ELCIC.wgee.single}} will calculate ELCIC value based on the function \code{\link{lambda.find.wgee.mean}} and \code{\link{ee.wgee.mean}}, which only involve estimating equations for marginal mean.
}
\examples{
## tests
# load data
data(wgeesimdata)
corstr<-"exchangeable"
dist<-"binomial"
x<-wgeesimdata$x
y<-wgeesimdata$y
x_mis<-wgeesimdata$x_mis
r<-wgeesimdata$obs_ind
id<-wgeesimdata$id
time<-3
index.var<-c(1,2,3)
ELCIC_value<-ELCIC.wgee.single(x,y,x_mis,r,id,time,index.var,name.var=NULL,
                     dist,corstr,joints=TRUE)
ELCIC_value
}
