% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geesimdata.R
\docType{data}
\name{geesimdata}
\alias{geesimdata}
\title{Data simulated for model selection under GEE framework without missingness}
\format{
An object of class \code{"list"}
\describe{
 \item{y}{The outcomes generated from Poisson distribution with three repeated measurements from each subject}
 \item{x}{A covariate matrix, of which the first column are all ones and rest columns contain normally distributed. Two are time-dependent variables, and one is time-independent variable.}
}
}
\usage{
data(geesimdata)
}
\description{
Data simulated for model selection under GEE framework without missingness
}
\examples{

data(geesimdata)
geesimdata$y

}
\references{
This data set was artificially created for the ELCIC package.
}
\keyword{datasets}
