% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_geetoydata.R
\docType{data}
\name{geetoydata}
\alias{geetoydata}
\title{Toydata generated for variable selection under GEE framework without missingness or missing completely at random}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{y}{The outcomes generated from Poisson distribution with three repeated measurements from each subject}
 \item{x}{A covariate matrix, of which the first column are all ones and rest columns contain normally distributed. Two are time-dependent variables, and one is time-independent variable.}
}
}
\usage{
data(geetoydata)
}
\description{
Toydata generated for variable selection under GEE framework without missingness or missing completely at random
}
\examples{

data(geetoydata)
geetoydata$y

}
\references{
This data set was artificially created for the ELCIC package.
}
\keyword{datasets}
