% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glmsimdata.R
\docType{data}
\name{glmsimdata}
\alias{glmsimdata}
\title{Data simulated for variable selection under GLM framework}
\format{
An object of class \code{"list"}
\describe{
 \item{y}{The outcome generated from Negative Binomial distribution with the dispersion parameter parameter=2}
 \item{x}{A covariate matrix, of which the first column are all ones and rest columns contain normally distributed
 values}
}
}
\usage{
data(glmsimdata)
}
\description{
Data simulated for variable selection under GLM framework
}
\examples{

data(glmsimdata)
glmsimdata$y

}
\references{
This data set was artificially created for the ELCIC package.
}
\keyword{datasets}
