% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_glmtoydata.R
\docType{data}
\name{glmtoydata}
\alias{glmtoydata}
\title{Toydata generated for variable selection under GLM framework}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{y}{The outcome generated from Negative Binomial distribution with the dispersion parameter parameter=2}
 \item{x}{A covariate matrix, of which the first column are all ones and rest columns contain normally distributed
 values}
}
}
\usage{
data(glmtoydata)
}
\description{
Toydata generated for variable selection under GLM framework
}
\examples{

data(glmtoydata)
glmtoydata$y

}
\references{
This data set was artificially created for the ELCIC package.
}
\keyword{datasets}
