% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{MLIC.wgee.procedure}
\alias{MLIC.wgee.procedure}
\title{The whole MLIC procedure for joint selection of mean structure and correlation structure in longitudinal data without missingness}
\usage{
MLIC.wgee.procedure(x,y,x_mis,r,id,candidate.sets,
      name.var.sets=NULL,dist,candidate.cor.sets, joints=TRUE)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes. use NA to indicate missing outcomes.}

\item{x_mis}{A matrix containing covariates for the missing data model. The first column should be all ones corresponding to the intercept.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records.}

\item{id}{A vector indicating subject id.}

\item{candidate.sets}{A list containing index corresponding to candidate covariates.}

\item{name.var.sets}{A list containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{candidate.cor.sets}{A vector containing condidate correlation structures. It is c("independence","exchangeable", "ar1") as default.}

\item{joints}{A logic value for joint selection of marginal mean and working correlation structure. The default is TRUE.}
}
\value{
A vector with each element containing MLIC value for each candidate model. The row name of this vector is the selected correlation structure.
}
\description{
This function provides the overall MLIC procedure for joint selection of mean structure and correlation structure in longitudinal data without missingness.
}
\examples{
## tests
# load data
data(wgeetoydata)
dist="binomial"
x=wgeetoydata$x
y=wgeetoydata$y
x_mis=wgeetoydata$x_mis
r=wgeetoydata$obs_ind
id=wgeetoydata$id
candidate.sets<-list(c(1,2),c(1,2,3))
candidate.cor.sets<-c("exchangeable","ar1")
criterion.mlic<-MLIC.wgee.procedure(x,y,x_mis,r,id,candidate.sets,
            name.var.sets=NULL,dist,candidate.cor.sets,joints=TRUE)
criterion.mlic
}
