% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{pii}
\alias{pii}
\title{Calculate the inverse of joint probability for observing records from each subject}
\usage{
pii(pi)
}
\arguments{
\item{pi}{A matrix containing covariates for the missing data model. The first column should be all ones corresponding to the intercept.}
}
\value{
pii: a vector containing the inverse of joint probabilities.
}
\description{
A function provides the inverse of joint probabilities for weight calculation involved in WGEE.
}
\examples{
## tests
# load data
data(wgeetoydata)
library(wgeesel)
data_wgee=data.frame(do.call(cbind,wgeetoydata))
corstr="exchangeable"
dist="binomial"
id=data_wgee$id
# obtain the estimates
fit=wgee(y~x1+x2+x3,data_wgee,id,family=dist,corstr =corstr,scale = NULL,
         mismodel =obs_ind~x_mis1+x_mis2)
beta=as.vector(summary(fit)$beta)
ro=summary(fit)$corr
phi=summary(fit)$phi
#calculate observing probabilies for all observations
gamma=as.vector(summary(fit$mis_fit)$coefficients[,1])
x_mis=wgeetoydata$x_mis
pi=prob.obs(x_mis,gamma)
joint_prob=pii(pi)
}
