% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supporting_functions.R
\name{prob.obs}
\alias{prob.obs}
\title{Calculate conditional probabilities for observing records from each subject}
\usage{
prob.obs(x_mis,gamma)
}
\arguments{
\item{x_mis}{A matrix containing covariates for the missing data model. The first column should be all ones corresponding to the intercept.}

\item{gamma}{coefficients calculated from missing data model}
}
\value{
pi: a vector containing conditional probabilities.
}
\description{
A function provides conditional probabilities for weight calculation in WGEE
}
\examples{
## tests
# load data
data(wgeetoydata)
library(wgeesel)
data_wgee=data.frame(do.call(cbind,wgeetoydata))
corstr="exchangeable"
dist="binomial"
id=data_wgee$id
# obtain the estimates
fit=wgee(y~x1+x2+x3,data_wgee,id,family=dist,corstr =corstr,scale = NULL,
         mismodel =obs_ind~x_mis1+x_mis2)
beta=as.vector(summary(fit)$beta)
ro=summary(fit)$corr
phi=summary(fit)$phi
#calculate observing probabilies for all observations
gamma<-as.vector(summary(fit$mis_fit)$coefficients[,1])
x_mis=wgeetoydata$x_mis
pi=prob.obs(x_mis,gamma)

}
