% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{QICc.procedure}
\alias{QICc.procedure}
\title{Joint selection of marginal mean and correlation structures in longitudinal data based on QIC}
\usage{
QICc.procedure(x,y,id,dist,candidate.sets=NULL, name.var.sets=NULL,
   candidate.cor.sets=c("independence","exchangeable", "ar1"), joints=TRUE)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{id}{A vector indicating subject id.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{candidate.sets}{A list containing index corresponding to candidate covariates.}

\item{name.var.sets}{A list containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{candidate.cor.sets}{A vector containing condidate correlation structures. When joint=TRUE, it is c("independence","exchangeable", "ar1") as default. When joint=FALSE, it should be one of c("independence","exchangeable", "ar1").}

\item{joints}{A logic value for joint selection of marginal mean and working correlation structure. The default is TRUE.}
}
\value{
A vector with each element containing QIC value for each candidate model. The row name of this vector is the selected correlation structure.
}
\description{
This function provides the Joint selection of marginal mean and correlation structures in longitudinal data based on QIC.
}
\details{
Either arguments "index.var" or "name.var" is used to identify the candidate mean model. If both arguments are provided, only the argument "name.var" will be used.

When the argument "joints" is TRUE, \code{\link{ELCIC.gee}} will calculate ELCIC value based on the function \code{\link{lambda.find.gee}} and \code{\link{ee.gee}}, which involve both marginal mean and correlation coefficient estimating equations. When the argument "joints" is FALSE, \code{\link{ELCIC.gee}} will calculate ELCIC value based on the function \code{\link{lambda.find.gee.onlymean}} and \code{\link{ee.gee.onlymean}}, which only involve marginal mean estimating equations.
}
\examples{
## tests
# load data
data(geetoydata)
x<-geetoydata$x
y<-geetoydata$y
id<-geetoydata$id
r<-rep(1,nrow(x))
time<-3
candidate.sets<-list(c(1,2),c(1,2,3))
candidate.cor.sets<-c("exchangeable")
dist="poisson"
criterion.qic<-QICc.procedure(x=x,y=y,id=id,dist=dist,candidate.sets=candidate.sets,
                    name.var.sets=NULL,candidate.cor.sets=candidate.cor.sets)
criterion.qic

}
