% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimating_equations.R
\name{ee.glm}
\alias{ee.glm}
\title{Estimating equations for ELCIC under GLM}
\usage{
ee.glm(x, y, betahat, dist)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{betahat}{A plug-in estimator solved by an external estimating procedure.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
A specified estimating equations for ELCIC under GLM.
}
\examples{
## tests
# load data
data(glmtoydata)
x=glmtoydata$x
y=glmtoydata$y
# obtain the estimates
fit<-glm(y~x-1,family="poisson")
betahat<-fit$coefficients
ee.matrix<-ee.glm(x, y, betahat, dist="poisson")
dim(ee.matrix)

}
