% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.gee.procedure}
\alias{ELCIC.gee.procedure}
\title{The whole procedure for joint selection of mean structure and correlation structure in longitudinal data without missingness}
\usage{
ELCIC.gee.procedure(x,y,r,id,time,candidate.sets=NULL,name.var.sets=NULL,dist,
      candidate.cor.sets=c("independence","exchangeable", "ar1"), joints=TRUE)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records. The default setup is that all data are observed.}

\item{id}{A vector indicating subject id.}

\item{time}{The number of observations for each subject}

\item{candidate.sets}{A list containing index corresponding to candidate covariates.}

\item{name.var.sets}{A list containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{candidate.cor.sets}{A vector containing condidate correlation structures. When joint=TRUE, it is c("independence","exchangeable", "ar1") as default. When joint=FALSE, it should be one of c("independence","exchangeable", "ar1").}

\item{joints}{A logic value for joint selection of marginal mean and working correlation structure. The default is TRUE.}
}
\value{
A matrix with each element containing ELCIC value for each candidate model.
}
\description{
This function provides the overall procedure for joint selection of mean structure and correlation structure in longitudinal data without missingness.
}
\examples{
## tests
# load data
data(geetoydata)
x=geetoydata$x
y=geetoydata$y
id=geetoydata$id
r=rep(1,nrow(x))
time=3
candidate.sets<-list(c(1,2),c(1,2,3))
candidate.cor.sets<-c("exchangeable")
dist="poisson"
criterion.elcic<-ELCIC.gee.procedure(x=x,y=y,r=r,id=id,time=time,candidate.sets=candidate.sets,
                name.var.sets=NULL,dist=dist,candidate.cor.sets=candidate.cor.sets)
criterion.elcic

}
