% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.gee}
\alias{ELCIC.gee}
\title{Calculate ELCIC value for a given candidate model under GEE framework without missingness}
\usage{
ELCIC.gee(x, y, r, id, time, index.var=NULL, name.var = NULL, dist, corstr, joints=TRUE)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records. The default setup is that all data are observed.}

\item{id}{A vector indicating subject id.}

\item{time}{The number of observations for each subject}

\item{index.var}{A vector containing index corresponding to candidate covariates.}

\item{name.var}{A vector containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{corstr}{A condidate correlation structure. It can be "independence","exchangeable", and "ar1".}

\item{joints}{A logic value for joint selection of marginal mean and working correlation structure. The default is TRUE.}
}
\value{
A ELCIC value for a given candidate model.
}
\description{
This is the function to calculate ELCIC value for a given candidate model. It is able to simultaneously evaluate mean model and working correlation structure.
}
\examples{
## tests
# load data
data(geetoydata)
x=geetoydata$x
y=geetoydata$y
id=geetoydata$id
r=rep(1,nrow(x))
time=3
corstr="exchangeable"
dist="poisson"
criteria<-ELCIC.gee(x=x,y=y,r=r,id=id,time=time,index.var=c(1,2,3),
           name.var=NULL,dist=dist,corstr=corstr)
criteria

}
