% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.glm.procedure}
\alias{ELCIC.glm.procedure}
\title{The whole variable selection procedure in GLM}
\usage{
ELCIC.glm.procedure(x,y,candidate.sets,name.var.sets=NULL,dist)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should contain all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{candidate.sets}{A list containing index corresponding to candidate covariates in each candidate model.}

\item{name.var.sets}{A list containing names of candidate covariates coresponding to each candidate model. The names should be subset of column names of the x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}
}
\value{
A matrix with each element containing ELCIC value for each candidate model (in columns) and (in rows)
}
\description{
This function provides the overall procedure for variable selection in GLM.
}
\examples{
## tests
# load data
data(glmtoydata)
x=glmtoydata$x
y=glmtoydata$y
#candidate model index
candidate.sets<-list(c(1,2),c(1,2,3),c(1,2,3,4))
criteria<-ELCIC.glm.procedure(x, y, candidate.sets, name.var.sets = NULL, dist="poisson")
criteria

}
