% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.glm}
\alias{ELCIC.glm}
\title{Variable selection in generalized linear models (GLMs)}
\usage{
ELCIC.glm(x, y, index.var, name.var = NULL, dist)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should contain all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{index.var}{A vector containing index corresponding to candidate covariates (including the intercept).}

\item{name.var}{A vector containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}
}
\value{
A vector containing information criteria including ELCIC, AIC, BIC, and GIC.
}
\description{
This function will provide values of several model selection criteria including AIC, BIC, GIC, and ELCIC.
}
\examples{
## tests
# load data
data(glmtoydata)
x=glmtoydata$x
y=glmtoydata$y
#candidate model index
name.var=c("intercept","x1","x2")
index.var=c(1,2,3)
criteria<-ELCIC.glm(x, y, index.var =index.var, name.var = NULL, dist="poisson")
criteria

}
