% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_selection.R
\name{ELCIC.wgee}
\alias{ELCIC.wgee}
\title{Calculate ELCIC value for a given candidate model under WGEE framework with data missing at random}
\usage{
ELCIC.wgee(x,y,x_mis,r,id,time,index.var,name.var=NULL,dist,corstr,joints=TRUE)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should be all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes. use NA to indicate missing outcomes.}

\item{x_mis}{A matrix containing covariates for the missing data model. The first column should be all ones corresponding to the intercept.}

\item{r}{A vector indicating missingness: 1 for observed records, and 0 for unobserved records.}

\item{id}{A vector indicating subject id.}

\item{time}{The number of observations for each subject}

\item{index.var}{A vector containing index corresponding to candidate covariates.}

\item{name.var}{A vector containing names of candidate covariates. The names should be subset of column names of x matrix.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}

\item{corstr}{A condidate correlation structure. It can be "independence","exchangeable", and "ar1".}

\item{joints}{A logic value for joint selection of marginal mean and working correlation structure. The default is TRUE.}
}
\value{
A matrix containing values of calculated estimating equations.
}
\description{
This is the function to calculate ELCIC value for a given candidate model. It is able to simultaneously evaluate mean model and working correlation structure. The data is missing at random.
}
\examples{
## tests
# load data
data(wgeetoydata)
corstr="exchangeable"
dist="binomial"
x=wgeetoydata$x
y=wgeetoydata$y
x_mis=wgeetoydata$x_mis
r=wgeetoydata$obs_ind
id=wgeetoydata$id
time=3
index.var=c(1,2,3)
ELCIC_value=ELCIC.wgee(x,y,x_mis,r,id,time,index.var,name.var=NULL,
                     dist,corstr,joints=TRUE)
}
