% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{lambda.find.glm}
\alias{lambda.find.glm}
\title{To calculate tuning parameter in the empirical likelihood}
\usage{
lambda.find.glm(x, y, betahat, dist)
}
\arguments{
\item{x}{A matrix containing covariates. The first column should contain all ones corresponding to the intercept.}

\item{y}{A vector containing outcomes.}

\item{betahat}{A plug-in estimator solved by an external estimating procedure.}

\item{dist}{A specified distribution. It can be "gaussian", "poisson",and "binomial".}
}
\value{
A value of lambda (tuning parameter) vector involved in the empirical likelihood.
}
\description{
This function aims to efficiently calculate the tuning parameter lambda in the empirical likelihood.
}
\examples{
## tests
# load data
data(glmtoydata)
x=glmtoydata$x
y=glmtoydata$y
# obtain the estimates
fit<-glm(y~x-1,family="poisson")
betahat<-fit$coefficients
lambda<-lambda.find.glm(x, y, betahat, dist="poisson")
lambda

}
